+++ dataworkspace/dataworkspace/apps/api_v1/urls.py
@@ -57,4 +57,11 @@ urlpatterns = [
             namespace="core",
         ),
     ),
+    path(
+        "data_insights/",
+        include(
+            ("dataworkspace.apps.api_v1.data_insights.urls", "data_insights"),
+            namespace="data_insights",
+        ),
+    ),
 ]
